More refactoring in anticipation of adding supercolumn support
