+++ libcassie/include.am
@@ -8,12 +8,14 @@ EXTRA_DIST+= \
 nobase_include_HEADERS+= \
 			 libcassie/cassie.h \
 			 libcassie/cassie_types.h \
+			 libcassie/cassie_io_column.h \
 			 libcassie/cassie_column.h
 
 lib_LTLIBRARIES+= libcassie/libcassie.la
 libcassie_libcassie_la_CXXFLAGS= ${AM_CXXFLAGS}
 libcassie_libcassie_la_SOURCES = \
 				       libcassie/cassie.cc \
+				       libcassie/cassie_io_column.cc \
 				       libcassie/cassie_column.cc
 
 libcassie_libcassie_la_DEPENDENCIES= libcassandra/libcassandra.la
